add confirm option to delete an user

+++ src/components/Users/Users.js
@@ -11,18 +11,21 @@ const Users = () => {
 
     // DELETE AN USER
     const handleDeleteUser = id => {
-        const url = `http://localhost:5000/users/${id}`;
-        fetch(url, {
-            method: 'DELETE'
-        })
-        .then(res => res.json())
-        .then(data => {
-            if(data.deletedCount > 0){
-                alert('deleted successfully!!');
-                const remainingUsers = users.filter(user => user._id !== id);
-                setUsers(remainingUsers);
-            }
-        })
+        const proceed = window.confirm('Are u sure, u want to delete?');
+        if(proceed){
+            const url = `http://localhost:5000/users/${id}`;
+            fetch(url, {
+                method: 'DELETE'
+            })
+            .then(res => res.json())
+            .then(data => {
+                if(data.deletedCount > 0){
+                    alert('deleted successfully!!');
+                    const remainingUsers = users.filter(user => user._id !== id);
+                    setUsers(remainingUsers);
+                }
+            });
+        }
     }
 
     return (
